requirements.txt
asgiref==3.8.1
attrs==25.3.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-django==4.11.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
